+++ src/typed_settings/converters.py
@@ -1,20 +1,35 @@
 """
-Converters and helpers for :mod:`cattrs`.
+Converters and structure hooks for various data types.
 """
 from datetime import datetime
 from enum import Enum
 from pathlib import Path
+from types import MappingProxyType
 from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
+    Dict,
+    Iterable,
     List,
+    Mapping,
     Optional,
     Protocol,
+    Sequence,
     Tuple,
     Type,
     Union,
+    get_args,
+    get_origin,
 )
+
+from ._compat import PY_310
+
+
+if PY_310:
+    from types import UnionType
+else:
+    from typing import Union as UnionType  # type: ignore
 
 import attrs
 import cattrs
@@ -26,61 +41,204 @@ from .types import ET, T
 
 if TYPE_CHECKING:
     import cattrs
-
-
-__all__ = [
-    "Converter",
-    "default_converter",
-    "get_default_structure_hooks",
-    "register_attrs_hook_factory",
-    "register_strlist_hook",
-    "to_dt",
-    "to_bool",
-    "to_enum",
-    "to_path",
-    "to_resolved_path",
-]
+#
+#
+# __all__ = [
+#     "Converter",
+#     "TSConverter",
+#     "default_converter",
+#     "get_default_structure_hooks",
+#     "register_attrs_hook_factory",
+#     "register_strlist_hook",
+#     "to_dt",
+#     "to_bool",
+#     "to_enum",
+#     "to_path",
+#     "to_resolved_path",
+# ]
 
 
 class Converter(Protocol):
-    def structure(self, data: Any, type: Type[T]) -> T:
+    """
+    **Protocol** that converters must implement.
+
+    Only a :meth:`structure()` method similar to the one from :program:`cattrs` is
+    required.
+
+    .. versionadded:: 23.1.0
+    """
+
+    def structure(self, value: Any, cls: Type[T]) -> T:
+        """
+        Convert *data* to an instance of *cls* and return it.
+
+        Args:
+            value: The data to be converted.
+            cls: The type to convert *data* to.
+
+        Return:
+            An instance of *cls* for *data*.
+        """
         ...
 
-    # def unstructure(self, inst: T) -> Any:
-    #     ...
+
+class TSConverter:
+    """
+    A simple converter that can replace :program:`cattrs` if you want to use
+    Typed Settings without dependencies.
+
+    It supports the same types as the default :program:`cattrs` converter.
+    """
+
+    def __init__(
+        self,
+        resolve_paths: bool = True,
+        strlist_sep: Union[str, Callable[[str], list], None] = ":",
+    ) -> None:
+        if strlist_sep is None:
+            self.strlist_hook: Optional[Callable[[str], list]] = None
+        elif isinstance(strlist_sep, str):
+            self.strlist_hook = lambda v: v.split(strlist_sep)  # type: ignore  # noqa
+        else:
+            self.strlist_hook = strlist_sep
+
+        self.scalar_converters: Dict[Any, Callable[[Any, type], Any]] = {
+            Any: to_any,
+            bool: to_bool,
+            int: to_type,
+            float: to_type,
+            str: to_type,
+            datetime: to_dt,
+            Enum: to_enum,
+            Path: to_resolved_path if resolve_paths else to_path,
+        }
+        self.composite_hook_factories: List[HookFactory] = [
+            ListHookFactory,
+            TupleHookFactory,
+            DictHookFactory,
+            MappingProxyTypeHookFactory,
+            SetHookFactory,
+            FrozenSetHookFactory,
+            UnionHookFactory,
+            AttrsHookFactory,
+        ]
+
+    def structure(self, value: Any, cls: Type[T]) -> T:
+        """
+        Recursively convert simple and composite types and return the result.
+
+        Args:
+            value: The value to convert
+        """
+        for ctype, convert in self.scalar_converters.items():
+            if cls is ctype or (
+                ctype is not Any and isinstance(cls, type) and issubclass(cls, ctype)
+            ):
+                return convert(value, cls)
+
+        origin = get_origin(cls)
+        args = get_args(cls)
+        for hook in self.composite_hook_factories:
+            if hook.match(cls, origin, args):
+                convert = hook.get_structure_hook(self, cls, origin, args)
+                return convert(value, cls)
+
+        raise TypeError(f"Cannot create converter for generic type: {cls}")
+
+    def maybe_apply_strlist_hook(self, value: T) -> Union[list, T]:
+        if self.strlist_hook and isinstance(value, str):
+            return self.strlist_hook(value)
+        return value
 
 
 def default_converter(*, resolve_paths: bool = True) -> Converter:
     """
-    Get an instanceof the default converter used by Typed Settings.
+    Get a default instances of a converter which will be used to convert/structured
+    the loaded settings.
 
     Args:
         resolve_path: Whether or not to resolve relative paths.
 
     Return:
-        A :class:`Converter` configured with addional hooks for loading the follwing
+        If :program:`cattrs` is installed, a :class:`cattrs.Converter`.  Else, a
+        :class:`TSConverter`.  The converters are configured to handle the following
         types:
 
-        - :class:`bool` using :func:`.to_bool()`
-        - :class:`datetime.datetime` using :func:`.to_dt()`
-        - :class:`enum.Enum` using :func:`.to_enum()`
-        - :class:`pathlib.Path`
+        - :class:`bool` (see :func:`to_bool()` for supported inputs)
+        - :class:`int`
+        - :class:`float`
+        - :class:`str`
+        - :class:`datetime.datetime` (see :func:`to_dt()`)
+        - :class:`enum.Enum` using (see :func:`to_enum()`)
+        - :class:`pathlib.Path` (see :func:`to_path()` and :func:`to_resolved_path()`)
+        - :class:`list`
+        - :class:`tuple`
+        - :class:`dict`
+        - :class:`types.MappingProxyType` ("read-only" dicts)
+        - :class:`set`
+        - :class:`frozenset`
+        - :data:`typing.Optional`
+        - :data:`typing.Union` (depending on the converter, only to a certain degree,
+          but this should not be relevant for settings with clearly defined types)
+        - :mod:`attrs` classes (from instances and dicts)
 
-        The converter can also structure attrs instances from existing attrs
-        instances (normaly, it would only work with dicts).  This allows using
-        instances of nested class es of default values for options.  See
-        :meth:`cattrs.converters.BaseConverter.register_structure_hook_factory()`.
+        :class:`list`, :class:`tuple`, :class:`set`, and :class:`frozenset` set can also
+        be converted from strings.  By default, strings are split by ``:``.  See
+        :class:`TSConverter` or :func:`register_strlist_hook()` for details.
 
     This converter can also be used as a base for converters with custom
     structure hooks.
+
+    .. versionchanged:: 23.1.0
+       Return a :program:`cattrs` converter if it is installed or else a Typed Settings
+       converter.
     """
-    return get_default_cattrs_converter(resolve_paths=resolve_paths)
+    try:
+        import cattrs  # noqa: F401
+    except ImportError:
+        return get_default_ts_converter(resolve_paths=resolve_paths)
+    else:
+        return get_default_cattrs_converter(resolve_paths=resolve_paths)
+
+
+def get_default_ts_converter(resolve_paths: bool = True) -> "TSConverter":
+    """
+    Return a :class:`TSConverter` with default settings
+    (see :func:`default_converter()` for argument and return value description).
+
+    Args:
+        resolve_path: Whether or not to resolve relative paths.
+
+    Return:
+        A :class:`TSConverter` instance with default configuration.
+    """
+    return TSConverter(resolve_paths=resolve_paths)
 
 
 def get_default_cattrs_converter(resolve_paths: bool = True) -> "cattrs.Converter":
-    import cattrs
+    """
+    Return a :class:`cattrs.Converter()` with default settings
+    (see :func:`default_converter()` for argument and return value description).
+
+    Args:
+        resolve_path: Whether or not to resolve relative paths.
+
+    Return:
+        A :class:`cattrs.Converter` instance with default configuration.
+
+    Raises:
+        ModuleNotFoundError: if :program:`cattrs` is not installed.
+    """
+    try:
+        import cattrs
+    except ImportError as e:
+        raise ModuleNotFoundError(
+            "Module 'cattrs' not installed.  Please run "
+            "'python -m pip install -U typed-settings[cattrs]'"
+        ) from e
 
     converter = cattrs.Converter()
+    register_mappingproxy_hook(converter)
     register_attrs_hook_factory(converter)
     register_strlist_hook(converter, ":")
     for t, h in get_default_structure_hooks(resolve_paths=resolve_paths):
@@ -97,6 +255,10 @@ def get_default_structure_hooks(
 
     Args:
         resolve_path: Whether or not to resolve relative paths.
+
+    Return:
+        A list of tuples that can be used as args for
+        :meth:`cattrs.Converter.register_structure_hook()`.
     """
     path_hook = to_resolved_path if resolve_paths else to_path
     return [
@@ -109,10 +271,13 @@ def get_default_structure_hooks(
 
 def register_attrs_hook_factory(converter: cattrs.Converter) -> None:
     """
-    Register a hook factory that allows using instances of attrs classes where
-    cattrs would normally expect a dictionary.
+    Register a hook factory that allows using instances of :program:`attrs` classes
+    where :program:`cattrs` would normally expect a dictionary.
 
     These instances are then returned as-is and without further processing.
+
+    Args:
+        converter: The :class:`cattrs.Converter` to register the hook at.
     """
 
     def allow_attrs_instances(typ):  # type: ignore[no-untyped-def]
@@ -126,6 +291,26 @@ def register_attrs_hook_factory(converter: cattrs.Converter) -> None:
     converter.register_structure_hook_factory(attrs.has, allow_attrs_instances)
 
 
+def register_mappingproxy_hook(converter: cattrs.Converter) -> None:
+    """
+    Register a hook factory for converting data to :class:`types.MappingProxyType`
+    instances.
+
+    Args:
+        converter: The :class:`cattrs.Converter` to register the hook at.
+    """
+
+    def check(cls: type) -> bool:
+        return cls is MappingProxyType or get_origin(cls) is MappingProxyType
+
+    def convert(val: Mapping, cls: Type[T]) -> T:
+        args = get_args(cls)
+        t = Dict[args[0], args[1]] if args else Dict  # type: ignore
+        return MappingProxyType(converter.structure(val, t))  # type: ignore
+
+    converter.register_structure_hook_func(check, convert)
+
+
 def register_strlist_hook(
     converter: cattrs.Converter,
     sep: Optional[str] = None,
@@ -137,7 +322,7 @@ def register_strlist_hook(
     environment variables).
 
     Args:
-        converter: The converter to register the hooks with.
+        converter: The :class:`cattrs.Converter` to register the hook at.
         sep: A separator used for splitting strings (see :meth:`str.split()`).
             Cannot be used together with *fn*.
         fn: A function that takes a string and returns a list, e.g.,
@@ -191,6 +376,13 @@ def _generate_hook_factory(structure_func, fn):  # type: ignore[no-untyped-def]
     return gen_func
 
 
+def to_any(value: Any, _cls: type) -> Any:
+    """
+    Return *value* as-is.
+    """
+    return value
+
+
 def to_bool(value: Any, _cls: type = bool) -> bool:
     """
     Convert "boolean" strings (e.g., from env. vars.) to real booleans.
@@ -198,22 +390,30 @@ def to_bool(value: Any, _cls: type = bool) -> bool:
     Values mapping to :code:`True`:
 
     - :code:`True`
-    - :code:`"true"` / :code:`"t"`
-    - :code:`"yes"` / :code:`"y"`
-    - :code:`"on"`
+    - :code:`"true"` / :code:`"t"` (case insensitive)
+    - :code:`"yes"` / :code:`"y"` (case insensitive)
+    - :code:`"on"` (case insensitive)
     - :code:`"1"`
     - :code:`1`
 
     Values mapping to :code:`False`:
 
     - :code:`False`
-    - :code:`"false"` / :code:`"f"`
-    - :code:`"no"` / :code:`"n"`
-    - :code:`"off"`
+    - :code:`"false"` / :code:`"f"` (case insensitive)
+    - :code:`"no"` / :code:`"n"` (case insensitive)
+    - :code:`"off"` (case insensitive)
     - :code:`"0"`
     - :code:`0`
 
-    Raise :exc:`ValueError` for any other value.
+    Args:
+        value: The value to parse.
+        _cls: (ignored)
+
+    Return:
+        A :class:`bool` for the input *value*.
+
+    Raise:
+        ValueError: If *value* is any other value than stated above.
     """
     if isinstance(value, str):
         value = value.lower()
@@ -241,7 +441,7 @@ def to_dt(value: Union[datetime, str], _cls: type = datetime) -> datetime:
 
     Args:
         value: The input data
-        _type: The desired output type, will be ignored
+        _cls: (ignored)
 
     Return:
         The converted datetime instance
@@ -262,8 +462,7 @@ def to_dt(value: Union[datetime, str], _cls: type = datetime) -> datetime:
 
 def to_enum(value: Any, cls: Type[ET]) -> ET:
     """
-    Return a converter that creates an instance of the :class:`~enum.Enum`
-    *cls*.
+    Return an instance of the enum *cls* for *value*.
 
     If the to be converted value is not already an enum, the converter will
     create one by name (``MyEnum[val]``).
@@ -286,11 +485,11 @@ def to_enum(value: Any, cls: Type[ET]) -> ET:
 
 def to_path(value: Union[Path, str], _cls: type) -> Path:
     """
-    Return *value* to :class:`~pathlib.Path`.
+    Convert *value* to :class:`~pathlib.Path`.
 
     Args:
         value: The input data
-        _cls: The path type, this argument is ignored.
+        _cls: (ignored)
 
     Return:
         An instance of :class:`~pathlib.Path`
@@ -303,16 +502,318 @@ def to_path(value: Union[Path, str], _cls: type) -> Path:
 
 def to_resolved_path(value: Union[Path, str], _cls: type) -> Path:
     """
-    Return *value* to :class:`.Path` and resolve it.
+    Convert *value* to :class:`.Path` and resolve it.
 
     Args:
         value: The input data
-        _cls: The path type, this argument is ignored.
+        _cls: (ignored)
 
     Return:
-        An instance of :class:`.Path`
+        A resolved instance of :class:`.Path`
 
     Raise:
         TypeError: If *value* cannot be converted to a path.
     """
     return Path(value).resolve()
+
+
+def to_type(value: Any, cls: Type[T]) -> T:
+    """
+    Convert *value* to *cls*.
+
+    Args:
+        value: The input data
+        cls: A class that takes a single argument, e.g., :class:`int`, :class:`float`,
+            or :class:`str`.
+
+    Return:
+        An instance of *cls*.
+
+    Raise:
+        ValueError: if *value* cannot be converted to *cls*.
+    """
+    return cls(value)  # type: ignore[call-arg]
+
+
+class HookFactory(Protocol):
+    """
+    **Protocol** for :class:`TSConverter` hook factories.
+
+    Hook factories have a :meth:`match` functions that decides whether they can handle
+    given type/class.  In addition, they can generate a structure hook for that type.
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        """
+        Check whether this class can handle the given type *cls*.
+
+        Args:
+            cls: The type/class to check.
+            origin: The type's origin as returned by :func:`typing.get_origin()`.
+            args: The type's args as retuned by :func:`typing.get_args()`.
+
+        Return:
+            ``True`` if this class can convert the given type, or else ``False``.
+        """
+        ...
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Any, Type[T]], T]:
+        """
+        Return a structure hook for the given type/class.
+
+        Args:
+            converter: The :class:`TSConverter` that the returned hook will be
+                registered at.  The structure hook can use the converter to recursively
+                convert sub elements of composite types.
+            cls: The type/class to convert to.
+            origin: The type's origin as returned by :func:`typing.get_origin()`.
+            args: The type's args as retuned by :func:`typing.get_args()`.
+
+        Return:
+            A structure hook, which is a function
+            :samp:`hook({value}: Any, {cls}: Type[T]) -> T`.
+        """
+        ...
+
+
+class AttrsHookFactory:
+    """
+    A :class:`HookFactory` that returns :program:`attrs` classes from dicts.  Instances
+    are of the given class are accepted as well and returned as-is (without further
+    processing of their attributes).
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return attrs.has(cls)
+
+    @staticmethod
+    def get_structure_hook(
+        converter: Converter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Union[dict, T], Type[T]], T]:
+        def convert(value: Union[dict, T], cls: Type[T]) -> T:
+            if isinstance(value, cls):
+                return value
+
+            if not isinstance(value, dict):
+                raise TypeError(
+                    f'Invalid type "{type(value).__name__}"; expected '
+                    f'"{cls.__name__}" or "dict".'
+                )
+
+            fields = attrs.fields_dict(cls)  # type: ignore[arg-type]
+            values = {
+                n: converter.structure(v, fields[n].type)  # type: ignore[arg-type]
+                for n, v in value.items()
+            }
+            return cls(**values)
+
+        return convert
+
+
+class ListHookFactory:
+    """
+    A :class:`HookFactory` for :class:`list` and :class:`typing.List`.
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return cls is list or origin is list
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Iterable, Type[T]], T]:
+        if not args:
+            args = (Any,)
+        item_type = args[0]
+
+        def convert(value: Iterable, cls: Type[T]) -> T:
+            value = converter.maybe_apply_strlist_hook(value)
+            values = [converter.structure(v, item_type) for v in value]
+            return list(values)  # type: ignore[return-value]
+
+        return convert
+
+
+class TupleHookFactory:
+    """
+    A :class:`HookFactory` for :class:`tuple` and :class:`typing.Tuple`.
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return cls is tuple or origin is tuple
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Union[Callable[[Iterable, Type[T]], T], Callable[[Sequence, Type[T]], T]]:
+        if not args:
+            args = (Any, ...)
+
+        convert: Union[
+            Callable[[Iterable, Type[T]], T],  # For list-like tuples
+            Callable[[Sequence, Type[T]], T],  # For struct-like tuples
+        ]
+        if len(args) == 2 and args[1] == ...:
+            item_type = args[0]
+
+            def convert(value: Iterable, cls: Type[T]) -> T:
+                value = converter.maybe_apply_strlist_hook(value)
+                values = [converter.structure(v, item_type) for v in value]
+                return tuple(values)  # type: ignore[return-value]
+
+        else:
+
+            def convert(value: Sequence, cls: Type[T]) -> T:
+                value = converter.maybe_apply_strlist_hook(value)
+                if len(value) != len(args):
+                    raise TypeError(
+                        f"Value must have {len(args)} items but has: {len(value)}"
+                    )
+                values = [converter.structure(v, t) for v, t in zip(value, args)]
+                return tuple(values)  # type: ignore[return-value]
+
+        return convert
+
+
+class DictHookFactory:
+    """
+    A :class:`HookFactory` for :class:`dict` and :class:`typing.Dict`.
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return cls is dict or origin is dict
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Mapping, Type[T]], T]:
+        if not args:
+            args = (Any, Any)
+        key_type, val_type = args
+
+        def convert(value: Mapping, cls: Type[T]) -> T:
+            values = {
+                converter.structure(k, key_type): converter.structure(v, val_type)
+                for k, v in value.items()
+            }
+            return values  # type: ignore[return-value]
+
+        return convert
+
+
+class MappingProxyTypeHookFactory:
+    """
+    A :class:`HookFactory` for :class:`types.MappingProxyType` (a read-only dict proxy).
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return cls is MappingProxyType or origin is MappingProxyType
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Any, Type[T]], T]:
+        if not args:
+            args = (Any, Any)
+        key_type, val_type = args
+
+        def convert(value: Mapping, cls: Type[T]) -> T:
+            values = {
+                converter.structure(k, key_type): converter.structure(v, val_type)
+                for k, v in value.items()
+            }
+            return MappingProxyType(values)  # type: ignore[return-value]
+
+        return convert
+
+
+class SetHookFactory:
+    """
+    A :class:`HookFactory` for :class:`set` and :class:`typing.Set`.
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return cls is set or origin is set
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Any, Type[T]], T]:
+        if not args:
+            args = (Any,)
+        item_type = args[0]
+
+        def convert(value: Iterable, cls: Type[T]) -> T:
+            value = converter.maybe_apply_strlist_hook(value)
+            values = [converter.structure(v, item_type) for v in value]
+            return set(values)  # type: ignore[return-value]
+
+        return convert
+
+
+class FrozenSetHookFactory:
+    """
+    A :class:`HookFactory` for :class:`frozenset` and :class:`typing.FrozenSet`.
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return cls is frozenset or origin is frozenset
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Any, Type[T]], T]:
+        if not args:
+            args = (Any,)
+        item_type = args[0]
+
+        def convert(value: Iterable, cls: Type[T]) -> T:
+            value = converter.maybe_apply_strlist_hook(value)
+            values = [converter.structure(v, item_type) for v in value]
+            return frozenset(values)  # type: ignore[return-value]
+
+        return convert
+
+
+class UnionHookFactory:
+    """
+    A :class:`HookFactory` for :data:`typing.Optional` and :data:`typing.Union`.
+
+    If the input data already has one of the uniton types, it will be returned without
+    further processing.  Otherwise, converters for all union types will be tried until
+    one works (i.e., raises no exception).
+    """
+
+    @staticmethod
+    def match(cls: type, origin: Optional[Any], args: Tuple[Any, ...]) -> bool:
+        return origin in (Union, UnionType)
+
+    @staticmethod
+    def get_structure_hook(
+        converter: TSConverter, cls: type, origin: Optional[Any], args: Tuple[Any, ...]
+    ) -> Callable[[Any, Type[T]], T]:
+        def convert(value: Any, cls: Type[T]) -> T:
+            if type(value) in args:
+                # Preserve val as-is if it already has a matching type.
+                # Otherwise float(3.2) would be converted to int
+                # if the converters are [int, float].
+                return value
+            for arg in args:
+                try:
+                    return converter.structure(value, arg)
+                except Exception:  # noqa: S110
+                    pass
+            raise ValueError(f"Failed to convert value to any Union type: {value}")
+
+        return convert